Clicking the button generates an activity.
The activites are displayed in the main activity now.

+++ app/src/main/java/com/example/myapplication_final/BoredData.kt
@@ -4,8 +4,8 @@ data class BoredData(
     val activity: String,
     val type: String,
     val participants: Int,
-    val price: Int,
+    val price: Float,
     val link: String,
     val key: String,
-    val accessibility: Int
+    val accessibility: Float
 )
